+++ src/RTree/LinearSplitRTreeNode.java
@@ -9,20 +9,20 @@ public class LinearSplitRTreeNode extends AbstractRTreeNode {
     public IRTreeNode[] split() {
         ArrayList<IRTreeNode> nodes = new ArrayList<IRTreeNode>();
         // Variables para determinar las separaciones
-        double maxLeft = Double.MIN_VALUE;
+        float maxLeft = Float.MIN_VALUE;
         int maxLeftIndex = 0;
-        double maxBottom = Double.MIN_VALUE;
+        float maxBottom = Float.MIN_VALUE;
         int maxBottomIndex = 0;
-        double minRight = Double.MAX_VALUE;
+        float minRight = Float.MAX_VALUE;
         int minRightIndex = 0;
-        double minTop = Double.MAX_VALUE;
+        float minTop = Float.MAX_VALUE;
         int minTopIndex = 0;
 
         // Variables para determinar la normalizacion
-        double minLeft = Double.MAX_VALUE;
-        double minBottom = Double.MAX_VALUE;
-        double maxRight = Double.MIN_VALUE;
-        double maxTop = Double.MIN_VALUE;
+        float minLeft = Float.MAX_VALUE;
+        float minBottom = Float.MAX_VALUE;
+        float maxRight = Float.MIN_VALUE;
+        float maxTop = Float.MIN_VALUE;
 
         for (int i = 0; i < size(); i++) {
             nodes.add(this.getChild(i));
@@ -57,12 +57,12 @@ public class LinearSplitRTreeNode extends AbstractRTreeNode {
             }
         }
 
-        double separationX = Math.abs(maxLeft - minRight);
-        double separationY = Math.abs(maxBottom - minTop);
-        double rangeX = maxRight - minLeft;
-        double rangeY = maxTop - minBottom;
-        double normalizedX = separationX / rangeX;
-        double normalizedY = separationY / rangeY;
+        float separationX = Math.abs(maxLeft - minRight);
+        float separationY = Math.abs(maxBottom - minTop);
+        float rangeX = maxRight - minLeft;
+        float rangeY = maxTop - minBottom;
+        float normalizedX = separationX / rangeX;
+        float normalizedY = separationY / rangeY;
 
         IRTreeNode initialNodeA;
         IRTreeNode initialNodeB;
